refactor: move GetWarnings into the alert module
Move GetWarnings from main.cpp into alert.{h,cpp}, next to the
cs_mapAlerts/mapAlerts storage it reads (issue #3125, workstream C6).
Pure code move, no behavior change.

main.cpp loses the function plus its ad-hoc extern declarations of
cs_mapAlerts/mapAlerts (alert.cpp defines them); the declaration moves
from main.h to alert.h, and the callers that relied on main.h for it
(rpc/mining.cpp, rpc/blockchain.cpp, wallet/rpcwallet.cpp) now include
alert.h directly -- qt/clientmodel.cpp and rpc/net.cpp already did.
