+++ src/main.cpp
@@ -1099,49 +1099,6 @@ void PrintBlockTree() EXCLUSIVE_LOCKS_REQUIRED(cs_main)
 
 //////////////////////////////////////////////////////////////////////////////
 //
-// CAlert
-//
-
-extern CCriticalSection cs_mapAlerts;
-extern map<uint256, CAlert> mapAlerts GUARDED_BY(cs_mapAlerts);
-
-string GetWarnings(string strFor)
-{
-    int nPriority = 0;
-    string strStatusBar;
-
-    // Misc warnings like out of disk space and clock is wrong
-    if (strMiscWarning != "")
-    {
-        nPriority = 1000;
-        strStatusBar = strMiscWarning;
-    }
-
-    // Alerts
-    {
-        LOCK(cs_mapAlerts);
-        for (auto const& item : mapAlerts)
-        {
-            const CAlert& alert = item.second;
-            if (alert.AppliesToMe() && alert.nPriority > nPriority)
-            {
-                nPriority = alert.nPriority;
-                strStatusBar = alert.strStatusBar;
-            }
-        }
-    }
-
-    if (strFor == "statusbar")
-    {
-        return strStatusBar;
-    }
-
-    assert(!"GetWarnings() : invalid parameter");
-    return "error";
-}
-
-//////////////////////////////////////////////////////////////////////////////
-//
 // Messages
 //
